ApplicationBot/requirements.txt
discord.py==2.5.2
gspread==6.1.2
python-dotenv==1.0.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0